Issue with getting the "highest number"
The script was looking for the "highest" number in the output. In this case, cause the Build number was the 16 and the Paper Version was the 18,  it was taking the "18" as the highest Number. I changed the script that it's deleting everything befor the build number until it's looking for the highest number in Output. Now its working again.